requirements.txt
Flask==3.0.0
python-dotenv>=1.0.0
twilio==8.10.0
google-generativeai>=0.3.0  # For Google Gemini API
# openai==1.3.7  # Commented out as using Google Gemini instead
python-dateutil==2.8.2
pytz==2023.3.post1
pytest==7.4.3
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
mypy==1.7.1
pyngrok>=7.2.0
# requirements.txt (add this line)
pydantic>=2.5.0
redis>=5.0.0